windows: Store the window handle in a pointer sized type
Under the hoods, HWND is a typedef to a void *, which is can't be stored
in the int in 64 bits systems.

+++ dali/internal/window-system/windows/window-base-win.h
@@ -363,7 +363,7 @@ private:
    * @param surface Any containing a surface id, or can be empty
    * @return surface id, or zero if surface is empty
    */
-  unsigned int GetSurfaceId( Any surface ) const;
+  uintptr_t GetSurfaceId( Any surface ) const;
 
   /**
    * @brief Create window
@@ -373,7 +373,7 @@ private:
   /**
    * @brief Sets up an already created window.
    */
-  void SetWinWindow( unsigned int surfaceId );
+  void SetWinWindow( uintptr_t surfaceId );
 
 private:
